move cache tests to a separate file
