+++ src/cache/filetree.rs
@@ -177,133 +177,3 @@ impl<'a> Iterator for Iter<'a> {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::cache::filetree::{EntryExistsError, FileTree};
-    use crate::types::{Directory, Entry, File};
-
-    fn sort_entries(entries: &mut Vec<Entry>) {
-        entries.sort_unstable_by(|e0, e1| e0.path().cmp(e1.path()));
-    }
-
-    fn example_tree_0() -> FileTree {
-        let mut filetree = FileTree::new();
-        assert_eq!(filetree.insert("a/0/x".into(), 1), Ok(()));
-        assert_eq!(filetree.insert("a/0/y".into(), 2), Ok(()));
-        assert_eq!(filetree.insert("a/1/x/0".into(), 7), Ok(()));
-        assert_eq!(filetree.insert("a/0/z/0".into(), 1), Ok(()));
-        assert_eq!(filetree.insert("a/1/x/1".into(), 2), Ok(()));
-        filetree
-    }
-
-    fn example_tree_1() -> FileTree {
-        let mut filetree = FileTree::new();
-        assert_eq!(filetree.insert("a/0/x".into(), 3), Ok(()));
-        assert_eq!(filetree.insert("a/0/y".into(), 2), Ok(()));
-        assert_eq!(filetree.insert("a/2/x/0".into(), 7), Ok(()));
-        assert_eq!(filetree.insert("a/0/z/0".into(), 9), Ok(()));
-        assert_eq!(filetree.insert("a/1/x/1".into(), 1), Ok(()));
-        filetree
-    }
-
-    fn example_tree_2() -> FileTree {
-        let mut filetree = FileTree::new();
-        assert_eq!(filetree.insert("b/0/x".into(), 3), Ok(()));
-        assert_eq!(filetree.insert("b/0/y".into(), 2), Ok(()));
-        assert_eq!(filetree.insert("a/2/x/0".into(), 7), Ok(()));
-        assert_eq!(filetree.insert("b/0/z/0".into(), 9), Ok(()));
-        assert_eq!(filetree.insert("a/1/x/1".into(), 1), Ok(()));
-        filetree
-    }
-
-    #[test]
-    fn insert_uniques_0() {
-        let mut entries = example_tree_0().iter().collect::<Vec<_>>();
-        sort_entries(&mut entries);
-        assert_eq!(entries, vec![
-            Entry::Directory(Directory { path: "a".into(), size: 13 }),
-            Entry::Directory(Directory { path: "a/0".into(), size: 4 }),
-            Entry::File(File { path: "a/0/x".into(), size: 1 }),
-            Entry::File(File { path: "a/0/y".into(), size: 2 }),
-            Entry::Directory(Directory { path: "a/0/z".into(), size: 1 }),
-            Entry::File(File { path: "a/0/z/0".into(), size: 1 }),
-            Entry::Directory(Directory { path: "a/1".into(), size: 9 }),
-            Entry::Directory(Directory { path: "a/1/x".into(), size: 9 }),
-            Entry::File(File { path: "a/1/x/0".into(), size: 7 }),
-            Entry::File(File { path: "a/1/x/1".into(), size: 2 }),
-        ]);
-    }
-
-    #[test]
-    fn insert_uniques_1() {
-        let mut entries = example_tree_1().iter().collect::<Vec<_>>();
-        sort_entries(&mut entries);
-        assert_eq!(entries, vec![
-            Entry::Directory(Directory { path: "a".into(), size: 22 }),
-            Entry::Directory(Directory { path: "a/0".into(), size: 14 }),
-            Entry::File(File { path: "a/0/x".into(), size: 3 }),
-            Entry::File(File { path: "a/0/y".into(), size: 2 }),
-            Entry::Directory(Directory { path: "a/0/z".into(), size: 9 }),
-            Entry::File(File { path: "a/0/z/0".into(), size: 9 }),
-            Entry::Directory(Directory { path: "a/1".into(), size: 1 }),
-            Entry::Directory(Directory { path: "a/1/x".into(), size: 1 }),
-            Entry::File(File { path: "a/1/x/1".into(), size: 1 }),
-            Entry::Directory(Directory { path: "a/2".into(), size: 7 }),
-            Entry::Directory(Directory { path: "a/2/x".into(), size: 7 }),
-            Entry::File(File { path: "a/2/x/0".into(), size: 7 }),
-        ]);
-    }
-
-    #[test]
-    fn insert_existing() {
-        let mut filetree = example_tree_0();
-        assert_eq!(filetree.insert("".into(), 1), Err(EntryExistsError));
-        assert_eq!(filetree.insert("a/0".into(), 1), Err(EntryExistsError));
-        assert_eq!(filetree.insert("a/0/z/0".into(), 1), Err(EntryExistsError));
-    }
-
-    #[test]
-    fn merge_test() {
-        let filetree = example_tree_0().merge(example_tree_1());
-        let mut entries = filetree.iter().collect::<Vec<_>>();
-        sort_entries(&mut entries);
-        assert_eq!(entries, vec![
-            Entry::Directory(Directory { path: "a".into(), size: 22 }),
-            Entry::Directory(Directory { path: "a/0".into(), size: 14 }),
-            Entry::File(File { path: "a/0/x".into(), size: 3 }),
-            Entry::File(File { path: "a/0/y".into(), size: 2 }),
-            Entry::Directory(Directory { path: "a/0/z".into(), size: 9 }),
-            Entry::File(File { path: "a/0/z/0".into(), size: 9 }),
-            Entry::Directory(Directory { path: "a/1".into(), size: 9 }),
-            Entry::Directory(Directory { path: "a/1/x".into(), size: 9 }),
-            Entry::File(File { path: "a/1/x/0".into(), size: 7 }),
-            Entry::File(File { path: "a/1/x/1".into(), size: 2 }),
-            Entry::Directory(Directory { path: "a/2".into(), size: 7 }),
-            Entry::Directory(Directory { path: "a/2/x".into(), size: 7 }),
-            Entry::File(File { path: "a/2/x/0".into(), size: 7 }),
-        ]);
-    }
-
-    #[test]
-    fn merge_reflexivity() {
-        assert_eq!(example_tree_0().merge(example_tree_0()), example_tree_0());
-        assert_eq!(example_tree_1().merge(example_tree_1()), example_tree_1());
-    }
-
-    #[test]
-    fn merge_associativity() {
-        assert_eq!(
-            example_tree_0().merge(example_tree_1()).merge(example_tree_2()),
-            example_tree_0().merge(example_tree_1().merge(example_tree_2()))
-        );
-    }
-
-    #[test]
-    fn merge_commutativity() {
-        assert_eq!(
-            example_tree_0().merge(example_tree_1()),
-            example_tree_1().merge(example_tree_0())
-        );
-    }
-}
